+++ sklearn/linear_model/__init__.py
@@ -16,7 +16,8 @@ from .least_angle import (Lars, LassoLars, lars_path, LarsCV, LassoLarsCV,
                           LassoLarsIC)
 from .coordinate_descent import (Lasso, ElasticNet, LassoCV, ElasticNetCV,
                                  lasso_path, enet_path, MultiTaskLasso,
-                                 MultiTaskElasticNet)
+                                 MultiTaskElasticNet, MultiTaskElasticNetCV,
+                                 MultiTaskLassoCV)
 from .sgd_fast import Hinge, Log, ModifiedHuber, SquaredLoss, Huber
 from .stochastic_gradient import SGDClassifier, SGDRegressor
 from .ridge import (Ridge, RidgeCV, RidgeClassifier, RidgeClassifierCV,
@@ -49,7 +50,9 @@ __all__ = ['ARDRegression',
            'LogisticRegression',
            'ModifiedHuber',
            'MultiTaskElasticNet',
+           'MultiTaskElasticNetCV',
            'MultiTaskLasso',
+           'MultiTaskLassoCV',
            'OrthogonalMatchingPursuit',
            'OrthogonalMatchingPursuitCV',
            'PassiveAggressiveClassifier',